feedback: merge cache and client
